programming/python/sqlalchemy-demo/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.25"
psycopg = "^3.1.17"
fastapi = "^0.108.0"
uvicorn = "^0.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
